Move most string util functions into string_utils.cpp from support.cpp
