+++ include/string_utils.h
@@ -137,8 +137,54 @@ void reset_str(T *str) noexcept
 	terminate_str_at(str, 0);
 }
 
-bool is_hex_digits(const std::string &s) noexcept;
+bool is_hex_digits(const std::string_view s) noexcept;
 
-bool is_digits(const std::string &s) noexcept;
+bool is_digits(const std::string_view s) noexcept;
+
+void strreplace(char *str, char o, char n);
+char *ltrim(char *str);
+char *rtrim(char *str);
+char *trim(char *str);
+char *upcase(char *str);
+char *lowcase(char *str);
+
+inline bool is_empty(const char *str) noexcept
+{
+	return str[0] == '\0';
+}
+
+bool IsHexWord(char *word);
+Bits ConvHexWord(char *word);
+char *StripWord(char *&cmd);
+
+std::string replace(const std::string &str, char old_char, char new_char) noexcept;
+void trim(std::string &str, const char trim_chars[] = " \r\t\f\n");
+void upcase(std::string &str);
+void lowcase(std::string &str);
+void strip_punctuation(std::string &str);
+
+// Split a string on an arbitrary character delimiter. Absent string content on
+// either side of a delimiter is treated as an empty string. For example:
+//   split("abc:", ':') returns {"abc", ""}
+//   split(":def", ':') returns {"", "def"}
+//   split(":", ':') returns {"", ""}
+//   split("::", ':') returns {"", "", ""}
+std::vector<std::string> split(const std::string &seq, const char delim);
+
+// Split a string on whitespace, where whitespace can be any of the following:
+// ' '    (0x20)  space (SPC)
+// '\t'   (0x09)  horizontal tab (TAB)
+// '\n'   (0x0a)  newline (LF)
+// '\v'   (0x0b)  vertical tab (VT)
+// '\f'   (0x0c)  feed (FF)
+// '\r'   (0x0d)  carriage return (CR)
+// Absent string content on either side of a delimiter is omitted. For example:
+//   split("abc") returns {"abc"}
+//   split("  a   b   c  ") returns {"a", "b", "c"}
+//   split("\t \n abc \r \v def \f \v ") returns {"abc", "def"}
+//   split("a\tb\nc\vd e\rf") returns {"a", "b", "c", "d", "e", "f"}
+//   split("  ") returns {}
+//   split(" ") returns {}
+std::vector<std::string> split(const std::string &seq);
 
 #endif
