+++ include/support.h
@@ -252,55 +252,10 @@ template <class T> T clamp(const T& n, const T& lower, const T& upper) {
 	return std::max<T>(lower, std::min<T>(n, upper));
 }
 
-void strreplace(char * str,char o,char n);
-char *ltrim(char *str);
-char *rtrim(char *str);
-char *trim(char * str);
-char * upcase(char * str);
-char * lowcase(char * str);
-
-inline bool is_empty(const char *str) noexcept
-{
-	return str[0] == '\0';
-}
 // Scans the provided command-line string for the '/'flag and removes it from
 // the string, returning if the flag was found and removed.
 bool ScanCMDBool(char *cmd, const char * flag);
 char * ScanCMDRemain(char * cmd);
-char * StripWord(char *&cmd);
-
-bool IsHexWord(char * word);
-Bits ConvHexWord(char * word);
-
-std::string replace(const std::string &str, char old_char, char new_char) noexcept;
-void trim(std::string &str, const char trim_chars[] = " \r\t\f\n");
-void upcase(std::string &str);
-void lowcase(std::string &str);
-void strip_punctuation(std::string &str);
-
-// Split a string on an arbitrary character delimiter. Absent string content on
-// either side of a delimiter is treated as an empty string. For example:
-//   split("abc:", ':') returns {"abc", ""}
-//   split(":def", ':') returns {"", "def"}
-//   split(":", ':') returns {"", ""}
-//   split("::", ':') returns {"", "", ""}
-std::vector<std::string> split(const std::string &seq, const char delim);
-
-// Split a string on whitespace, where whitespace can be any of the following:
-// ' '    (0x20)  space (SPC)
-// '\t'   (0x09)  horizontal tab (TAB)
-// '\n'   (0x0a)  newline (LF)
-// '\v'   (0x0b)  vertical tab (VT)
-// '\f'   (0x0c)  feed (FF)
-// '\r'   (0x0d)  carriage return (CR)
-// Absent string content on either side of a delimiter is omitted. For example:
-//   split("abc") returns {"abc"}
-//   split("  a   b   c  ") returns {"a", "b", "c"}
-//   split("\t \n abc \r \v def \f \v ") returns {"abc", "def"}
-//   split("a\tb\nc\vd e\rf") returns {"a", "b", "c", "d", "e", "f"}
-//   split("  ") returns {}
-//   split(" ") returns {}
-std::vector<std::string> split(const std::string &seq);
 
 bool is_executable_filename(const std::string &filename) noexcept;
