+++ tests/support_tests.cpp
@@ -40,125 +40,6 @@ TEST(DriveIndex, DriveZ)
 	EXPECT_EQ(25, drive_index('Z'));
 }
 
-TEST(Support_split_delim, NoBoundingDelims)
-{
-	const std::vector<std::string> expected({"a", "/b", "/c/d", "/e/f/"});
-	EXPECT_EQ(split("a:/b:/c/d:/e/f/", ':'), expected);
-	EXPECT_EQ(split("a /b /c/d /e/f/", ' '), expected);
-	EXPECT_EQ(split("abc", 'x'), std::vector<std::string>{"abc"});
-}
-
-TEST(Support_split_delim, DelimAtStartNotEnd)
-{
-	const std::vector<std::string> expected({"", "a", "/b", "/c/d", "/e/f/"});
-	EXPECT_EQ(split(":a:/b:/c/d:/e/f/", ':'), expected);
-	EXPECT_EQ(split(" a /b /c/d /e/f/", ' '), expected);
-}
-
-TEST(Support_split_delim, DelimAtEndNotStart)
-{
-	const std::vector<std::string> expected({"a", "/b", "/c/d", "/e/f/", ""});
-	EXPECT_EQ(split("a:/b:/c/d:/e/f/:", ':'), expected);
-	EXPECT_EQ(split("a /b /c/d /e/f/ ", ' '), expected);
-}
-
-TEST(Support_split_delim, DelimsAtBoth)
-{
-	const std::vector<std::string> expected({"", "a", "/b", "/c/d", "/e/f/", ""});
-	EXPECT_EQ(split(":a:/b:/c/d:/e/f/:", ':'), expected);
-	EXPECT_EQ(split(" a /b /c/d /e/f/ ", ' '), expected);
-}
-
-TEST(Support_split_delim, MultiInternalDelims)
-{
-	const std::vector<std::string> expected(
-	        {"a", "/b", "", "/c/d", "", "", "/e/f/"});
-	EXPECT_EQ(split("a:/b::/c/d:::/e/f/", ':'), expected);
-	EXPECT_EQ(split("a /b  /c/d   /e/f/", ' '), expected);
-}
-
-TEST(Support_split_delim, MultiBoundingDelims)
-{
-	const std::vector<std::string> expected(
-	        {"", "", "a", "/b", "/c/d", "/e/f/", "", "", ""});
-	EXPECT_EQ(split("::a:/b:/c/d:/e/f/:::", ':'), expected);
-	EXPECT_EQ(split("  a /b /c/d /e/f/   ", ' '), expected);
-}
-
-TEST(Support_split_delim, MixedDelims)
-{
-	const std::vector<std::string> expected(
-	        {"", "", "a", "/b", "", "/c/d", "/e/f/"});
-	EXPECT_EQ(split("::a:/b::/c/d:/e/f/", ':'), expected);
-	EXPECT_EQ(split("  a /b  /c/d /e/f/", ' '), expected);
-}
-
-TEST(Support_split_delim, Empty)
-{
-	const std::vector<std::string> empty;
-	const std::vector<std::string> two({"", ""});
-	const std::vector<std::string> three({"", "", ""});
-
-	EXPECT_EQ(split("", ':'), empty);
-	EXPECT_EQ(split(":", ':'), two);
-	EXPECT_EQ(split("::", ':'), three);
-	EXPECT_EQ(split("", ' '), empty);
-	EXPECT_EQ(split(" ", ' '), two);
-	EXPECT_EQ(split("  ", ' '), three);
-}
- 
-TEST(Support_split, NoBoundingWhitespace)
-{
-	const std::vector<std::string> expected({"a", "/b", "/c/d", "/e/f/"});
-	EXPECT_EQ(split("a /b /c/d /e/f/"), expected);
-	EXPECT_EQ(split("abc"), std::vector<std::string>{"abc"});
-}
-TEST(Support_split, WhitespaceAtStartNotEnd)
-{
-	const std::vector<std::string> expected({"a", "/b", "/c/d", "/e/f/"});
-	EXPECT_EQ(split(" a /b /c/d /e/f/"), expected);
-}
-
-TEST(Support_split, WhitespaceAtEndNotStart)
-{
-	const std::vector<std::string> expected({"a", "/b", "/c/d", "/e/f/"});
-	EXPECT_EQ(split("a /b /c/d /e/f/ "), expected);
-}
-
-TEST(Support_split, WhitespaceAtBoth)
-{
-	const std::vector<std::string> expected({"a", "/b", "/c/d", "/e/f/"});
-	EXPECT_EQ(split(" a /b /c/d /e/f/ "), expected);
-}
-
-TEST(Support_split, MultiInternalWhitespace)
-{
-	const std::vector<std::string> expected({"a", "/b", "/c/d", "/e/f/"});
-	EXPECT_EQ(split("a /b  /c/d   /e/f/"), expected);
-}
-
-TEST(Support_split, MultiBoundingWhitespace)
-{
-	const std::vector<std::string> expected({"a", "/b", "/c/d", "/e/f/"});
-	EXPECT_EQ(split("  a /b /c/d /e/f/   "), expected);
-}
-
-TEST(Support_split, MixedWhitespace)
-{
-	const std::vector<std::string> expected({"a", "b", "c"});
-	EXPECT_EQ(split("\t\na\f\vb\rc"), expected);
-	EXPECT_EQ(split("a\tb\f\vc"), expected);
-	EXPECT_EQ(split(" a \n \v \r b \f \r c "), expected);
-}
-
-TEST(Support_split, Empty)
-{
-	const std::vector<std::string> empty;
-	EXPECT_EQ(split(""), empty);
-	EXPECT_EQ(split(" "), empty);
-	EXPECT_EQ(split("   "), empty);
-}
-
 TEST(Support_next_int, Signed)
 {
 	// 8-bit types should upgrade to int16_t
