+++ tests/vs/tests.vcxproj.filters
@@ -6,6 +6,7 @@
     <ClCompile Include="..\..\src\misc\fs_utils_win32.cpp" />
     <ClCompile Include="..\..\src\misc\rwqueue.cpp" />
     <ClCompile Include="..\..\src\misc\setup.cpp" />
+    <ClCompile Include="..\..\src\misc\string_utils.cpp" />
     <ClCompile Include="..\..\src\misc\support.cpp" />
     <ClCompile Include="..\..\src\libs\ghc\fs_std_impl.cpp" />
     <ClCompile Include="..\..\src\libs\loguru\loguru.cpp" />
